Adding a specific error for 482 "auth. commands issued out of sequence"
Reporting 482 "authentication commands issued out of sequence" as
AUTHENTIFICATION REJECTED tricks clients into thinking that the
supplied credentials are invalid, while it is definitely not the case
(but rather, this is caused by a flaw in the client's logic).

AFAIK there is no adequate error for this in libetpan, hence the
introduction of a new error:
  NEWSNNTP_ERROR_AUTHENTICATION_OUT_OF_SEQUENCE

It is mapped to MAIL_ERROR_BAD_STATE in nntpdriver-tools:
nntpdriver_nntp_error_to_mail_error() --not sure about this, but
included here so that it is not forgotten and it will be eventually
fixed in case the patch is accepted.

Signed-off-by: Sébastien Bigaret <sebastien.bigaret@telecom-bretagne.eu>