Adding OAuth2 authentication support
Implements two functions:
mailimap_oauth2_authenticate: Use to to send the authentication token
to the server and actually authenticate with it.
mailimap_oauth2_authenticate_send Builds and sends the authentication
string in the form of base64("user=" {User} "^Aauth=Bearer " {Access
Token} "^A^A")`
This has been tested against Gmail.