Support Xcode in non-standard locations.
Use xcode-select --print-path to find the currently selected Xcode.
This avoids hardcoding /Applications/Xcode.app, so that it is
possible to use Xcode from non-standard locations (I have Xcode4
and Xcode5 installed in parallel).

Note that it is not possible to have Xcode installed in a path
with a space (I used "Xcode 5" originally).  I tried to fix this
but it looks like sasl's configure and build would never manage it.