Fix IMAP ID send.
Use mailimap_quoted_send rather than mailimap_astring_send when
sending names and values inside mailimap_id_param_send.
This means skipping the "workaround for buggy Courier-IMAP that
does not accept quoted-strings for fields name but prefer atoms"
and instead always sending quoted strings.

RFC 2971 (with reference to RFC 2060) is clear that these fields
may only be quoted strings or literals, not atoms.  Yahoo's server
rejects this command when atoms are sent (and it is correct to do so).

We shouldn't break compat with well-behaving servers as a workaround
for buggy ones, so using quoted strings here is the best choice.